Refactor: update imports for new decorators
