+++ agenta-cli/agenta/__init__.py
@@ -1,5 +1,4 @@
 from .sdk.utils.preinit import PreInitObject
-from .sdk.agenta_decorator import app, entrypoint
 from .sdk.context import get_contexts, save_context
 from .sdk.types import (
     Context,
@@ -14,7 +13,8 @@ from .sdk.types import (
     FileInputURL,
     BinaryParam,
 )
-from .sdk.tracing.decorators import span
+from .sdk.decorators.tracing import span
+from .sdk.decorators.llm_entrypoint import entrypoint
 from .sdk.agenta_init import Config, init, llm_tracing
 from .sdk.utils.helper.openai_cost import calculate_token_usage
