+++ agenta-cli/agenta/sdk/__init__.py
@@ -1,6 +1,4 @@
 from .utils.preinit import PreInitObject  # always the first import!
-from . import agenta_decorator, context, types, utils  # noqa: F401
-from .agenta_decorator import app, entrypoint
 from .context import get_contexts, save_context
 from .types import (
     Context,
@@ -15,7 +13,8 @@ from .types import (
     FileInputURL,
     BinaryParam,
 )
-from .tracing.decorators import span
+from .decorators.tracing import span
+from .decorators.llm_entrypoint import entrypoint
 from .agenta_init import Config, init, llm_tracing
 from .utils.helper.openai_cost import calculate_token_usage
